pyproject.toml
[tool.poetry]
name = "p"
version = "0.1.0"
description = ""
authors = ["Hecjin <atiliogoulart@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
pre-commit = "^2.15.0"
drf-yasg = "^1.21.3"
django-environ = "^0.9.0"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
